requirements.txt
websockets==12.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
python-telegram-bot==21.0.1
flask==3.0.3
